Release/v3.0.3

+++ Library/AviasalesSDK.framework/Headers/JRSDKPriceBuilder.h
@@ -3,7 +3,7 @@
 @interface JRSDKPriceBuilder : NSObject
 
 @property (strong, nonatomic, nullable) JRSDKCurrency currency;
-@property (assign, nonatomic) NSInteger value;
+@property (assign, nonatomic) float value;
 
 - (BOOL)canBuild;
 - (nullable JRSDKPrice *)build;
